feat: refactor post service into explore and management
