+++ src/posts/posts.controller.ts
@@ -11,7 +11,6 @@ import {
   UseGuards,
   UseInterceptors,
 } from '@nestjs/common';
-import { PostsService } from './posts.service';
 import { CreatePostDto } from './dto/request/create-post.dto';
 import { PostDetailsResponseDto } from './dto/response/post-details.dto';
 import { UpdatePostDto } from './dto/request/update-post.dto';
@@ -21,23 +20,33 @@ import { CurrentUser } from 'src/auth/decorators/users.decorator';
 import { CurrentUserType } from 'src/auth/types/current-user.type';
 import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
 import { Public } from 'src/auth/decorators/public.decorator';
+import { PostsManagementService } from './posts-management.service';
+import { PostsExploreService } from './posts-explore.service';
 
-@UseGuards(JwtAuthGuard)
 @Controller('posts')
 export class PostsController {
-  constructor(private readonly postsService: PostsService) {}
+  constructor(
+    private readonly postsManagementService: PostsManagementService,
+    private readonly postsExploreService: PostsExploreService,
+  ) {}
 
   @Post()
+  @UseGuards(JwtAuthGuard)
   @UseInterceptors(FilesInterceptor('images'))
   async createPost(
     @Body() createPostDto: CreatePostDto,
     @UploadedFiles() images: Express.Multer.File[],
     @CurrentUser() user: CurrentUserType,
   ): Promise<any> {
-    return this.postsService.createPost(createPostDto, images, user.id);
+    return this.postsManagementService.createPost(
+      createPostDto,
+      images,
+      user.id,
+    );
   }
 
   @Patch(':post_id')
+  @UseGuards(JwtAuthGuard)
   @UseInterceptors(FilesInterceptor('images'))
   async updatePost(
     @Param('post_id') postId: number,
@@ -45,7 +54,7 @@ export class PostsController {
     @UploadedFiles() images: Express.Multer.File[],
     @CurrentUser() user: CurrentUserType,
   ): Promise<PostDetailsResponseDto> {
-    return this.postsService.updatePost(
+    return this.postsManagementService.updatePost(
       Number(postId),
       updatePostDto,
       images,
@@ -54,8 +63,9 @@ export class PostsController {
   }
 
   @Delete(':post_id')
+  @UseGuards(JwtAuthGuard)
   async deletePost(@Param('post_id') postId: number) {
-    return this.postsService.deletePost(Number(postId));
+    return this.postsManagementService.deletePost(Number(postId));
   }
 
   @Get('search')
@@ -64,7 +74,7 @@ export class PostsController {
     @Query('page') page: string = '1', // using string here because somehow default value is not working with number
     @Query('page_size') page_size: string = '25',
   ): Promise<PostListItemResponseDto[]> {
-    return this.postsService.searchPosts(
+    return this.postsExploreService.searchPosts(
       query,
       Number(page),
       Number(page_size),
@@ -72,23 +82,30 @@ export class PostsController {
   }
 
   @Post('for-you')
+  @UseGuards(JwtAuthGuard)
   async getForYouPosts(
     @Body() body: { page: number; page_size: number; filter: string[] },
     @CurrentUser() user: CurrentUserType,
   ): Promise<PostListItemResponseDto[]> {
     const { page = 1, page_size = 25, filter } = body;
 
-    return this.postsService.getForYouPosts(user.id, page, page_size, filter);
+    return this.postsExploreService.getForYouPosts(
+      user.id,
+      page,
+      page_size,
+      filter,
+    );
   }
 
   @Post('following')
+  @UseGuards(JwtAuthGuard)
   async getFollowingPosts(
     @Body() body: { page: number; page_size: number; filter: string[] },
     @CurrentUser() user: CurrentUserType,
   ): Promise<PostListItemResponseDto[]> {
     const { page = 1, page_size = 24, filter } = body;
 
-    return this.postsService.getFollowingPosts(
+    return this.postsExploreService.getFollowingPosts(
       user.id,
       page,
       page_size,
@@ -101,7 +118,7 @@ export class PostsController {
   async getPostDetails(
     @Param('post_id') postId: number,
   ): Promise<PostDetailsResponseDto> {
-    return this.postsService.getPostDetails(Number(postId));
+    return this.postsExploreService.getPostDetails(Number(postId));
   }
 
   @Get('user/:username')
@@ -110,7 +127,7 @@ export class PostsController {
     @Query('page') page: string = '1',
     @Query('page_size') pageSize: string = '25',
   ): Promise<PostListItemResponseDto[]> {
-    return this.postsService.findPostsByUsername(
+    return this.postsExploreService.findPostsByUsername(
       username,
       Number(page),
       Number(pageSize),
